Extract per-sender email stats query into shared helper
